third_party/blink/web_tests/external/wpt/trusted-types/trusted-types-svg-script-expected.txt

This is a testharness.js-based test.
[FAIL] Assign String to SVGScriptElement.innerHTML.
  assert_equals: expected "SVGScriptElement innerHTML|'modified via innerHTML';" but got "Element innerHTML|'modified via innerHTML';"
[FAIL] Assign TrustedHTML to SVGScriptElement.innerHTML.
  assert_equals: a single violation reported expected 1 but got 0
[FAIL] Assign TrustedHTML to SVGScriptElement.innerHTML and execute it.
  assert_equals: a single violation reported expected 1 but got 0
[FAIL] Modify SVGScriptElement via DOM manipulation.
  assert_equals: a single violation reported expected 1 but got 0
Harness: the test ran to completion.
